+++ src/components/internalview/projectsmessageboard/pmGrid.jsx
@@ -6,14 +6,18 @@ const PmGrid = () => (
     <Header as='h3' dividing>
       Comments
     </Header>
+    {/* dyanmically generated project title */}
 
     <Comment>
+      {/* dynamically generated from comment
+      avatar=icon of user
+      author fName + Lname
+      Text 
+     */}
       <Comment.Avatar src='/images/avatar/small/matt.jpg' />
       <Comment.Content>
         <Comment.Author as='a'>Matt</Comment.Author>
-        <Comment.Metadata>
-          <div>Today at 5:42PM</div>
-        </Comment.Metadata>
+        
         <Comment.Text>How artistic!</Comment.Text>
         <Comment.Actions>
           <Comment.Action>Reply</Comment.Action>
@@ -25,9 +29,7 @@ const PmGrid = () => (
       <Comment.Avatar src='/images/avatar/small/elliot.jpg' />
       <Comment.Content>
         <Comment.Author as='a'>Elliot Fu</Comment.Author>
-        <Comment.Metadata>
-          <div>Yesterday at 12:30AM</div>
-        </Comment.Metadata>
+      
         <Comment.Text>
           <p>This has been very useful for my research. Thanks as well!</p>
         </Comment.Text>
@@ -40,9 +42,6 @@ const PmGrid = () => (
           <Comment.Avatar src='/images/avatar/small/jenny.jpg' />
           <Comment.Content>
             <Comment.Author as='a'>Jenny Hess</Comment.Author>
-            <Comment.Metadata>
-              <div>Just now</div>
-            </Comment.Metadata>
             <Comment.Text>Elliot you are always so right :)</Comment.Text>
             <Comment.Actions>
               <Comment.Action>Reply</Comment.Action>
@@ -56,9 +55,6 @@ const PmGrid = () => (
       <Comment.Avatar src='/images/avatar/small/joe.jpg' />
       <Comment.Content>
         <Comment.Author as='a'>Joe Henderson</Comment.Author>
-        <Comment.Metadata>
-          <div>5 days ago</div>
-        </Comment.Metadata>
         <Comment.Text>Dude, this is awesome. Thanks so much</Comment.Text>
         <Comment.Actions>
           <Comment.Action>Reply</Comment.Action>
