+++ src/components/internalview/projectsmessageboard/projectmessageboard.jsx
@@ -1,14 +1,14 @@
 import React from "react"
 import PmGrid from "./pmGrid"
 import { Container } from "semantic-ui-react"
-import Header from "../header/header"
-import Footer from "../footer/footer"
+import IHeader from "../header/iheader"
+import IFooter from "../footer/ifooter"
 function IHomepage() {
   return (
     <Container>
-      <Header />
+      <IHeader />
       <PmGrid />
-      <Footer />
+      <IFooter />
     </Container>
      
   )
